refactor: (#175) 타입 분리 수정
